refactor(tests): move mockOperator to shared test file
